New Command Added
Added a new command to give the player a custom item that has been loaded into item manager. Use 'give {item manager name} {desired amount}' to use this command. I also had to rename 'Honey 1' to 'newHoney' in the asset bundle as the space in the name was splicing the space as an arg in the command. Just something to keep in mind for future items.